+++ packages/figma/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @ryndesign/figma
 
+## 0.1.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @ryndesign/core@0.1.2
+
 ## 0.1.1
 
 ### Patch Changes